+++ CHANGELOG.md
@@ -6,11 +6,18 @@ Use spec: https://common-changelog.org/
 
 ### Changed
 
-- Upgraded to `go1.25`
-- Removed charm's `wish` with pico's `pssh`
+## v2026-02-25
+
+### Changed
+
+- Replaced `--comment` flag which was a string into a bool and now require comment to be provided by stdin for commands `accept`, `close`, and `reopen`
+  - `echo "lgtm!" | ssh pr.pico.sh pr accept --comment 100`
+  - If no `--comment` flag provided then you don't need to provide stdin
 
 ## v2026-02-24
 
 ### Changed
 
 - Added `ssh {username}@pr register` command and now require explicit registration to use this service
+- Upgraded to `go1.25`
+- Removed charm's `wish` with pico's `pssh`